Fix: Fix race issue with empty search query
Prior to this commit, if the search field was erased too quickly, the
results of the empty query would be populated into the state before the
results of the previous query had completed. This caused search results
for the first letter of the previous query (the last letter erased) to
show with an empty query.